Removed the PatchLabel target and contents.
