--- PatchLabel/PatchLabel.h
@@ -1,19 +1,0 @@
-//
-//  PatchLabel.h
-//  PatchLabel
-//
-//  Created by Kevin McKee on 10/11/17.
-//  Copyright © 2017 Procore. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//! Project version number for PatchLabel.
-FOUNDATION_EXPORT double PatchLabelVersionNumber;
-
-//! Project version string for PatchLabel.
-FOUNDATION_EXPORT const unsigned char PatchLabelVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <PatchLabel/PublicHeader.h>
-
-
